+++ packages/cater-build/plugins/favicon.js
@@ -1,0 +1,3 @@
+// Copyright Jon Williams 2017-2018. See LICENSE file.
+
+module.exports = function plugin() {};