+++ packages/cater-build/src/plugins.js
@@ -1,26 +1,71 @@
 // Copyright Jon Williams 2017-2018. See LICENSE file.
+const fs = require('fs');
 const path = require('path');
 
 /*
  * Loads plugin modules defined in the cater options.
  */
 
-function autoDefinePlugins(pkg, defaultPlugins) {
+function autoDefinePlugins(appRootPath, pkg, defaultPlugins) {
+  const plugins = {};
+
+  // Define from packages
   const deps = Object.keys(pkg.dependencies || {});
   const dev = Object.keys(pkg.devDependencies || {});
-  const plugins = {};
 
   const configuredDefaults = defaultPlugins
     .filter((entry) => Array.isArray(entry))
     .map(([name]) => name);
+  const caterDeps = deps.concat(dev).filter((k) => k.startsWith('cater-'));
 
-  deps
-    .concat(dev)
-    .filter((k) => k.startsWith('cater-'))
+  // Packages with plugin as the root element
+  caterDeps
     .filter((k) => require.resolve(k).endsWith('plugin.js'))
     .filter((k) => !configuredDefaults.includes(k))
     .forEach((k) => {
       plugins[k] = null;
+    });
+
+  // Packages with a plugins directory - this will take any cater-* packages
+  // and look for a "plugins" directory in the package root. All *.js files
+  // in that directory (but not subdirectories) are turned into a package.
+  // The current app root is included too.
+  caterDeps
+    .reduce(
+      // Turns a list of packages into a list of root directories for each
+      // package
+      // ['cater-build', 'cater-jest'] => ['/Code/node_modules/cater-build', ...]
+      (prev, curr) => {
+        let dir = path.dirname(require.resolve(curr));
+        while (!fs.existsSync(path.join(dir, 'package.json'))) {
+          const next = path.join(dir, '..');
+          if (dir === next) return prev; // Hit the root directory
+          dir = next;
+        }
+        prev.push(dir);
+        return prev;
+      },
+      [appRootPath]
+    )
+    // Reduces + transforms the package directories to those that have a
+    // plugins directory
+    .reduce((prev, curr) => {
+      const dir = path.join(curr, 'plugins');
+      if (fs.existsSync(dir)) {
+        const stat = fs.statSync(dir);
+        if (stat.isDirectory()) prev.push(dir);
+      }
+      return prev;
+    }, [])
+    .reduce((prev, curr) => {
+      const files = fs.readdirSync(curr).filter((v) => v.match(/\.js$/));
+      const paths = files
+        .map((name) => path.join(curr, name))
+        .filter((filename) => fs.statSync(filename).isFile());
+      return prev.concat(paths);
+    }, [])
+    .forEach((dir) => {
+      plugins[dir] = null;
     });
 
   return plugins;
@@ -34,6 +79,12 @@ function configurePlugin(cater, name, options) {
     );
   }
   const plugin = require(name); // eslint-disable-line global-require, import/no-dynamic-require
+
+  if (typeof plugin !== 'function') {
+    throw new Error(
+      `Cater plugin ${name} was specified, but that module did not export a function. Plugins should export a 'function(cater, options)' as the default.`
+    );
+  }
 
   // Plugin should export a function to configure itself
   plugin.result = plugin(cater, options);
@@ -57,7 +108,7 @@ function configure(app, options) {
   let plugins = [];
   if (options.plugins === 'auto') {
     const pkg = app.package();
-    plugins = autoDefinePlugins(pkg, options.defaultPlugins);
+    plugins = autoDefinePlugins(app.appRootPath, pkg, options.defaultPlugins);
   }
 
   return configurePlugins(app, plugins, options);