+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Included a plugin to enable babel-transformed use of the polished CSS-in-JS utility library. Use cater-polished to add support.
 * Added in spread operators via the babel transformation.
+* Simple plugins can now be created by dropping a _.js file in the /plugins directory of your app, or in any of the included cater-_ packages.
 
 ### Fixed